Don't crash if built docs are not available.

+++ src/weecfg/station_config.py
@@ -541,13 +541,19 @@ def copy_docs(config_dict, docs_root=None, dry_run=False, force=False):
     if os.path.isdir(docs_dir) and not force:
         print(f"Directory {docs_dir} already exists. Nothing done.")
         return None
-    print(f"Removing {docs_dir}.")
-    if not dry_run:
-        shutil.rmtree(docs_dir, ignore_errors=True)
     with weeutil.weeutil.get_resource_path('weewx_data', 'docs') as docs_resources:
-        print(f"Copying new docs into {docs_dir}.")
-        if not dry_run:
-            shutil.copytree(docs_resources, docs_dir)
+        # Check whether new docs are actually available.
+        if os.path.isdir(docs_resources):
+            # They are available. Proceed.
+            print(f"Removing {docs_dir}.")
+            if not dry_run:
+                shutil.rmtree(docs_dir, ignore_errors=True)
+            print(f"Copying new docs into {docs_dir}.")
+            if not dry_run:
+                shutil.copytree(docs_resources, docs_dir)
+        else:
+            print(f"{bcolors.WARNING}Documentation not available at {docs_resources}. "
+                  f"Try building them using mkdocs.{bcolors.ENDC}", file=sys.stderr)
     return docs_dir
